refactor #step2 유저 이름 리팩토링
