+++ README.md
@@ -24,7 +24,7 @@
 |-----|-----| 모양과 같이 가로 라인이 겹치는 경우 어느 방향으로 이동할지 결정할 수 없다.
 ```
 1. 사다리 초기화
-- [ ] 사다리 게임에 참여하는 사람의 이름을 받는다.
+- [x] 사다리 게임에 참여하는 사람의 이름을 받는다.
     -[x] 이름은 5자이상 일 시 exception
     -[x] 쉼표(,)를 구분자로 한다.
 - [ ] 사다리를 생성한다.
